solve 2b
add the dampener and handle the cases that this introduces.
the picking the element to remove is somewhat naive, it assumes it's the
current, if that doesn't work it's the next. after having removed both, the resulting
series should always be invalid without removing additonal elements.

rewrite tests to use assert_eq as the ouput is quicker to parse if something goes wrong